Add files via upload
Code for the menu and the game arena. If you want to add anything to it do it in the Game() loop, not the menu one since it's all done already. The code is well commented and easy to use. I am going to continue developing it for the next 2 days until the deadline.